jobs/etl_jobs/workflows/titelive/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==25.1.0
    # via gcloud-aio-storage
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   -r requirements.in
    #   gcloud-aio-auth
aiosignal==1.4.0
    # via aiohttp
anyio==4.12.1
    # via httpx
asyncio==4.0.0
    # via -r requirements.in
attrs==25.4.0
    # via aiohttp
backoff==2.2.1
    # via
    #   -r requirements.in
    #   gcloud-aio-auth
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
chardet==5.2.0
    # via gcloud-aio-auth
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via typer
cryptography==46.0.3
    # via gcloud-aio-auth
db-dtypes==1.5.0
    # via pandas-gbq
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gcloud-aio-auth==5.4.2
    # via gcloud-aio-storage
gcloud-aio-storage==9.6.1
    # via -r requirements.in
google-api-core==2.29.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
    #   pandas-gbq
google-auth==2.47.0
    # via
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==1.2.2
    # via
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.40.0
    # via
    #   -r requirements.in
    #   pandas-gbq
google-cloud-core==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-secret-manager==2.26.0
    # via -r requirements.in
google-cloud-storage==3.7.0
    # via -r requirements.in
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.3
    # via google-cloud-secret-manager
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
numpy==2.4.1
    # via
    #   db-dtypes
    #   pandas
    #   pandas-gbq
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
pandas==2.3.3
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
pandas-gbq==0.33.0
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.27.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==6.33.3
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psutil==7.2.1
    # via pandas-gbq
pyarrow==22.0.0
    # via
    #   -r requirements.in
    #   db-dtypes
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via
    #   gcloud-aio-storage
    #   google-auth
pycparser==2.23
    # via cffi
pydata-google-auth==1.9.1
    # via pandas-gbq
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via gcloud-aio-auth
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2025.2
    # via pandas
requests==2.32.5
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.2.0
    # via typer
rsa==4.9.1
    # via
    #   gcloud-aio-storage
    #   google-auth
setuptools==80.9.0
    # via
    #   pandas-gbq
    #   pydata-google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
tqdm==4.67.1
    # via -r requirements.in
typer==0.21.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   grpcio
    #   typer
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
yarl==1.22.0
    # via aiohttp
